+++ src/catch2/catch_session.cpp
@@ -187,6 +187,7 @@ namespace Catch {
             showHelp();
         if( m_configData.libIdentify )
             libIdentify();
+
         m_config.reset();
         return 0;
     }
@@ -255,6 +256,7 @@ namespace Catch {
         if (m_configData.showHelp || m_configData.libIdentify) {
             return 0;
         }
+
 
         CATCH_TRY {
             config(); // Force config to be constructed